+++ packages/orq-rc/src/orq_ai_sdk/models/findonebykeypromptsnippetop.py
@@ -152,6 +152,9 @@ r"""The version of photoReal to use. Must be v1 or v2. Only available for `leona
 FindOneByKeyPromptSnippetEncodingFormat = Literal["float", "base64"]
 r"""The format to return the embeddings"""
 
+FindOneByKeyPromptSnippetReasoningEffort = Literal["low", "medium", "high"]
+r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
+
 
 class FindOneByKeyPromptSnippetModelParametersTypedDict(TypedDict):
     r"""Model Parameters: Not all parameters apply to every model"""
@@ -195,6 +198,8 @@ class FindOneByKeyPromptSnippetModelParametersTypedDict(TypedDict):
     r"""The version of photoReal to use. Must be v1 or v2. Only available for `leonardoai` provider"""
     encoding_format: NotRequired[FindOneByKeyPromptSnippetEncodingFormat]
     r"""The format to return the embeddings"""
+    reasoning_effort: NotRequired[FindOneByKeyPromptSnippetReasoningEffort]
+    r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
 
 
 class FindOneByKeyPromptSnippetModelParameters(BaseModel):
@@ -264,6 +269,12 @@ class FindOneByKeyPromptSnippetModelParameters(BaseModel):
     encoding_format: Optional[FindOneByKeyPromptSnippetEncodingFormat] = None
     r"""The format to return the embeddings"""
 
+    reasoning_effort: Annotated[
+        Optional[FindOneByKeyPromptSnippetReasoningEffort],
+        pydantic.Field(alias="reasoningEffort"),
+    ] = None
+    r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = [
@@ -282,6 +293,7 @@ class FindOneByKeyPromptSnippetModelParameters(BaseModel):
             "responseFormat",
             "photoRealVersion",
             "encoding_format",
+            "reasoningEffort",
         ]
         nullable_fields = ["responseFormat"]
         null_default_fields = []
@@ -710,6 +722,11 @@ r"""The version of photoReal to use. Must be v1 or v2. Only available for `leona
 FindOneByKeyPromptSnippetPromptSnippetsEncodingFormat = Literal["float", "base64"]
 r"""The format to return the embeddings"""
 
+FindOneByKeyPromptSnippetPromptSnippetsReasoningEffort = Literal[
+    "low", "medium", "high"
+]
+r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
+
 
 class FindOneByKeyPromptSnippetPromptSnippetsModelParametersTypedDict(TypedDict):
     r"""Model Parameters: Not all parameters apply to every model"""
@@ -755,6 +772,10 @@ class FindOneByKeyPromptSnippetPromptSnippetsModelParametersTypedDict(TypedDict)
     r"""The version of photoReal to use. Must be v1 or v2. Only available for `leonardoai` provider"""
     encoding_format: NotRequired[FindOneByKeyPromptSnippetPromptSnippetsEncodingFormat]
     r"""The format to return the embeddings"""
+    reasoning_effort: NotRequired[
+        FindOneByKeyPromptSnippetPromptSnippetsReasoningEffort
+    ]
+    r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
 
 
 class FindOneByKeyPromptSnippetPromptSnippetsModelParameters(BaseModel):
@@ -827,6 +848,12 @@ class FindOneByKeyPromptSnippetPromptSnippetsModelParameters(BaseModel):
     )
     r"""The format to return the embeddings"""
 
+    reasoning_effort: Annotated[
+        Optional[FindOneByKeyPromptSnippetPromptSnippetsReasoningEffort],
+        pydantic.Field(alias="reasoningEffort"),
+    ] = None
+    r"""Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response."""
+
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = [
@@ -845,6 +872,7 @@ class FindOneByKeyPromptSnippetPromptSnippetsModelParameters(BaseModel):
             "responseFormat",
             "photoRealVersion",
             "encoding_format",
+            "reasoningEffort",
         ]
         nullable_fields = ["responseFormat"]
         null_default_fields = []
@@ -1301,7 +1329,7 @@ class FindOneByKeyPromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-01-31T13:31:12.935Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-02T16:01:07.474Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")